requirements.txt
opencv-python
numpyPillow
python-dotenv
langchain
langchain-google-genai
deepgram-sdk
pyaudio
google-generativeai